+++ examples/paper/conf.py
@@ -13,27 +13,33 @@ cddm.conf.set_verbose(2)
 #: width and height of the frame
 SIZE = 512
 #: how many frames to simulate
-NFRAMES = 1024*4
+NFRAMES = 1024
 #: max k in the first axis
-KIMAX = 31
+KIMAX = 39
 #: max k in the second axis
-KJMAX = 31
+KJMAX = 39
 #: the delta parameter for the simulator
 DELTA = 1
 #: image static background value
-BACKGROUND = 16384
+BACKGROUND = 2**14
 #: peak intensity of the particles
-INTENSITY = 512
+INTENSITY = 2**9
 #: sigma of the gaussian of the particles
 SIGMA = 3
 #: simulation video margin size
-MARGIN = 32
+MARGIN = 0
 #: period of the random triggering
-PERIOD = 32
+PERIOD = 8
+#:how many particles to simulate
+NUM_PARTICLES = 100
 
-NFRAMES_FULL = NFRAMES * int(PERIOD**0.5)
+BIT_DEPTH = "12bit"
 
-NPARTICLES = 100
+NOISE_MODEL = "gaussian"
+
+READOUT_NOISE = 0.
+
+SATURATION = 2**15
 
 #computed parameters, do not change these
 #----------------------------------------
@@ -49,6 +55,23 @@ KJSIZE = KJMAX+1
 
 #: diffusion constant
 D = 0.5*(DELTA*2*np.pi/SIZE)**2
+
+N_PARAMETER = PERIOD//2
+
+NFRAMES_FULL = NFRAMES * int(PERIOD**0.5)
+
+
+NFRAMES_RANDOM = NFRAMES
+
+NFRAMES_STANDARD = NFRAMES*2
+
+NFRAMES_DUAL = NFRAMES
+
+#: clipping value of the image (max value)
+VMAX = 2**int(BIT_DEPTH.split("bit")[0])
+#: mean image value
+VMEAN = BACKGROUND / SATURATION * VMAX
+