+++ examples/paper/cross_correlate.py
@@ -1,37 +1,33 @@
 """
-Demonstrates how to compute fft of videos and the compute auto correlation
-function with the out-of-memory version of the multitau algorithm.
+Demonstrates how to compute fft of videos and the compute cross-correlation
+function with the out-of-memory version of the multitau algorithm and do
+live view of the computation.
 """
-from cddm.viewer import DataViewer, CorrViewer
-from cddm.video import multiply, normalize_video, crop, asarrays
+from cddm.viewer import CorrViewer
+from cddm.video import multiply, normalize_video, crop, load, asarrays
 from cddm.window import blackman
 from cddm.fft import rfft2, normalize_fft
-from cddm.core import acorr, normalize, stats
+from cddm.core import ccorr, normalize,stats
 from cddm.multitau import log_average
-from cddm.sim import seed
 
 import numpy as np
-
-seed(0)
+from examples.paper.conf import  PERIOD, SHAPE, KIMAX, KJMAX, NFRAMES_RANDOM, DATA_PATH
 
 #: see video_simulator for details, loads sample video
-import examples.paper.random_video as video_simulator
+import dual_video
 import importlib
+importlib.reload(dual_video) #recreates iterator
 
-
-importlib.reload(video_simulator) #recreates iterator
-
-from examples.paper.conf import KIMAX, KJMAX, SHAPE, NFRAMES, DATA_PATH
-
+t1, t2 = dual_video.t1, dual_video.t2 
 
 #: create window for multiplication...
 window = blackman(SHAPE)
 
 #: we must create a video of windows for multiplication
-window_video = ((window,),)*NFRAMES
+window_video = ((window,window),)*NFRAMES_RANDOM
 
 #:perform the actual multiplication
-video = multiply(video_simulator.video, window_video)
+video = multiply(dual_video.video, window_video)
 
 #: if the intesity of light source flickers you can normalize each frame to the intensity of the frame
 #video = normalize_video(video)
@@ -39,21 +35,15 @@ video = multiply(video_simulator.video, window_video)
 #: perform rfft2 and crop results, to take only first kimax and first kjmax wavenumbers.
 fft = rfft2(video, kimax = KIMAX, kjmax = KJMAX)
 
-#: you can also normalize each frame with respect to the [0,0] component of the fft
-#: this it therefore equivalent to  normalize_video
-#fft = normalize_fft(fft)
-
 #load in numpy array
-fft_array, = asarrays(fft, NFRAMES)
+fft1,fft2 = asarrays(fft, NFRAMES_RANDOM)
 
 if __name__ == "__main__":
     import os.path as p
 
     #: now perform auto correlation calculation with default parameters 
-    data = acorr(fft_array, t = video_simulator.t, n = NFRAMES)
-    bg, var = stats(fft_array)
-    
-    
+    data = ccorr(fft1,fft2, t1 = t1,t2 = t2, n = NFRAMES_RANDOM)
+    bg, var = stats(fft1,fft2)
     
     #: perform normalization and merge data
     data_lin = normalize(data, bg, var, scale = True)
@@ -69,7 +59,5 @@ if __name__ == "__main__":
     x,y = log_average(data_lin, size = 16)
     
     #: save the normalized data to numpy files
-    np.save(p.join(DATA_PATH, "corr_random_t.npy"),x)
-    np.save(p.join(DATA_PATH, "corr_random_data.npy"),y)
-
-
+    np.save(p.join(DATA_PATH, "corr_dual_t.npy"),x)
+    np.save(p.join(DATA_PATH, "corr_dual_data.npy"),y)