+++ examples/paper/standard_video.py
@@ -3,14 +3,15 @@
 from cddm.sim import simple_brownian_video, seed, adc
 from cddm.viewer import VideoViewer 
 from cddm.video import load, crop, multiply
-from examples.paper.conf import NFRAMES, SIMSHAPE, BACKGROUND, DELTA, INTENSITY, SIGMA, SHAPE,DUST1_PATH
+from examples.paper.conf import NFRAMES_STANDARD, SIMSHAPE, BACKGROUND, DELTA, \
+    INTENSITY, PERIOD, SIGMA, SHAPE,DUST1_PATH, BIT_DEPTH, VMAX, NOISE_MODEL, SATURATION, READOUT_NOISE
 import matplotlib.pyplot as plt
 
 import numpy as np
 
 #: this cretaes a brownian motion frame iterator. 
 #: each element of the iterator is a tuple holding a single numpy array (frame)
-video = simple_brownian_video(range(NFRAMES), shape = SIMSHAPE,background = BACKGROUND,
+video = simple_brownian_video(range(NFRAMES_STANDARD), shape = SIMSHAPE,background = BACKGROUND, dt = PERIOD/2.,
                               sigma = SIGMA, delta = DELTA, intensity = INTENSITY, dtype = "uint16")
 
 
@@ -19,11 +20,11 @@ video = crop(video, roi = ((0,SHAPE[0]), (0,SHAPE[1])))
 
 #: apply dust particles
 dust = plt.imread(DUST1_PATH)[...,0] #float normalized to (0,1)
-dust = ((dust,),)*NFRAMES
+dust = ((dust,),)*NFRAMES_STANDARD
 
 video = multiply(video, dust)
 
-video = (tuple((adc(f, bit_depth = "12bit") for f in frames)) for frames in video)
+video = (tuple((adc(f, noise_model = NOISE_MODEL, saturation = SATURATION, readout_noise = READOUT_NOISE, bit_depth = BIT_DEPTH) for f in frames)) for frames in video)
 
 
 if __name__ == "__main__":
@@ -32,5 +33,5 @@ if __name__ == "__main__":
     #video = load(video, NFRAMES) # loads and displays progress bar
 
     #: VideoViewer either expects a multi_frame iterator, or a numpy array
-    viewer = VideoViewer(video, count = NFRAMES, vmin = 0, cmap = "gray", vmax = 4096)
+    viewer = VideoViewer(video, count = NFRAMES_STANDARD, vmin = 0, cmap = "gray", vmax = VMAX)
     viewer.show()